how to save date 'day-month-year' :D

+++ app/src/main/java/com/example/duan1_vinh/model/KhoanThu.java
@@ -3,23 +3,23 @@ package com.example.duan1_vinh.model;
 import java.util.Date;
 
 public class KhoanThu {
-    private int sotien;
+    private double sotien;
     private String ghichu;
     private LoaiThu loaiThu;
     private Date ngaygio;
 
-    public KhoanThu(int sotien, String ghichu, LoaiThu loaiThu, Date ngaygio) {
+    public KhoanThu(double sotien, String ghichu, LoaiThu loaiThu, Date ngaygio) {
         this.sotien = sotien;
         this.ghichu = ghichu;
         this.loaiThu = loaiThu;
         this.ngaygio = ngaygio;
     }
 
-    public int getSotien() {
+    public double getSotien() {
         return sotien;
     }
 
-    public void setSotien(int sotien) {
+    public void setSotien(double sotien) {
         this.sotien = sotien;
     }
